# backend/docs/swagger/openapi.yml
openapi: 3.0.3
info:
  title: ERP Distribution - API Gateway
  version: 1.0.0
  description: |
    Specification OpenAPI centralisee exposee via l'API Gateway.
    Les clients consomment principalement ces routes /api/v1/*.
servers:
  - url: http://localhost:4000
    description: API Gateway locale
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: erp_session
paths:
...
  /api/v1/sales/orders:
    get:
      summary: Lister les commandes (via gateway)
      tags: [Sales]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Liste des commandes
        '403':
          description: Permission orders.read requise
    post:
      summary: Creer une commande (via gateway)
      tags: [Sales]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: string
                  description: Optionnel (auto-genere si absent, ex OR-2026-000001)
                customer_id:
                  type: string
                customer_name:
                  type: string
                  description: Alternative a customer_id
                order_date:
                  type: string
                  format: date
                ship_date:
                  type: string
                  format: date
                lines:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                      product_name:
                        type: string
                        description: Alternative a product_id
                      quantity:
                        type: integer
                      discount:
                        type: number
                      unit_price:
                        type: number
                      cost:
                        type: number
                      sales:
                        type: number
                      profit:
                        type: number
              required: [order_date, lines]
      responses:
        '201':
          description: Commande creee
        '409':
          description: Periode comptable cloturee
        '422':
          description: Validation metier echouee (dates/lignes/montants)
        '403':
          description: Permission orders.create requise
  /api/v1/sales/orders/{orderId}:
    get:
      summary: Detail commande (via gateway)
      tags: [Sales]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail commande
        '404':
          description: Non trouve
        '403':
          description: Permission orders.read requise
    patch:
      summary: Mettre a jour une commande (via gateway)
      tags: [Sales]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Commande mise a jour
        '409':
          description: Periode comptable cloturee
        '422':
          description: Validation metier echouee
        '403':
          description: Permission orders.update requise
    delete:
      summary: Supprimer une commande (via gateway)
      tags: [Sales]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Commande supprimee
        '409':
          description: Periode comptable cloturee
        '403':
          description: Permission orders.delete requise
  /api/v1/sales/orders/{orderId}/transition:
    post:
      summary: Transitionner le statut d'une commande (workflow controle)
      tags: [Sales]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to_status:
                  type: string
                  enum: [Draft, Submitted, Approved, Shipped, Closed]
              required: [to_status]
      responses:
        '200':
          description: Transition appliquee
        '404':
          description: Commande non trouvee
        '409':
          description: Periode comptable cloturee
        '422':
          description: Transition invalide ou statut invalide
        '403':
          description: Permission orders.transition requise
...